Don't pass zero-size arrays to qsort(); they may be NULL.

+++ file.c
@@ -31,9 +31,10 @@ void tag_init (tag_t * tag, const char * name)
 {
     changeset_init (&tag->changeset);
     tag->changeset.type = ct_tag;
-    tag->changeset.time = -1l << (sizeof (time_t) * 8 - 1);
+    tag->changeset.time = -(1ul << (sizeof (time_t) * 8 - 1));
     assert (tag->changeset.time < 0);
-    assert ((tag->changeset.time & (tag->changeset.time - 1)) == 0);
+    assert ((tag->changeset.time & ((unsigned long) tag->changeset.time - 1))
+            == 0);
 
     tag->tag = name;
     tag->tag_files = NULL;
